feat: add almost working solution, not appended correctly
I have almost all I need, prior to refactoring. The problem at the
moment is that paths with project.el come with the absolute path, which
removes the headache, but I have to then split the path string into:

1. absolute path without file name, and
2. file name.

The second is easy enough, and will be the only part visible in the
minibuffer calling consult--multi. For this, I use the same approach as
consult-projectile: just make the remaining part invisible.