feat: add updated file path and project file source
The problem was described in the previous commit. The fix was as
follows. Given that project-files returns a list with the absolute
paths (interesting, and better), split said file paths in (directory
until file name, file name).

We make the first part (too long to display) invisible and we
concatenate it with the second part.