fix: fix relative root path on project files, change command name
Previously, the project file list was computed hiding the absolute path
up to the file name, but this lead to not knowing where the file is in
the directory tree.

Now, all the file candidates are displayed with their paths relative to
the project root.

BREAKING CHANGE:
Rename the consult-project-extra function to consult-project, see #2.