fix: fix project source not being called
Seems weird, but having a correct modification of the buffer source
fixed the project source issue, and therefore the third issue.

There is still work to be done with this first command, i.e, having a
correct behaviour when the command is called outside any
projects. consult--project is supposed to help with this, but the
project-current function is not returning the select project, and makes
you select again (in this case by only showing the project source).

Closes #3.