fix: remove consult-project-extra buffer source
I don't really need to use a custom buffer source, and less so for the
reasons I stated in the comment on the previous commits. The
consult--source-project-buffer source already uses ?b as narrow key, and
the :hidden problem is gone.

It seems to have been another error with the code. Can't point to it atm.