feat: add project-find-file as default
Checking whether the project is current outside the function might not
be the best, and some performance issues are raised, but it creates a
foolproof method for reaching the files in every situation.

Here is an example of the two use cases I have:
1. Run consult-project-extra-find upon startup to find a file in a project.
2. Run consult-project-extra-find when already on a project, to move
quickly between files in the project and known projects.

With this change, 1. is supported via the project-find-file function,
while the second is using our custom sources. We are also closing #4, as
the case isn't reached anymore.

Closes #4.