feat: add elpa project dependency
It seemed that project.el didn't have the project-known-roots function
in Emacs 27.1, which is the Emacs version I added. Just add the project
elpa dependency, which is still fine.

Currently I'm trying to add the package to Melpa, but it shouldn't be a
problem to have it accepted on GNU Elpa either, as we are only depending
on Emacs and GNU Elpa.