Recommend extending project-prefix-map in README
I would recommend binding the additional commands to `project-prefix-map`, which
in my version of emacs 29 is by default bound to `C-x p` and not `C-c p`, which
was used in the README example.

Using the project prefix map for all project commands is useful because it makes
remembering and discovering commands easier, e.g. with `which-key` I just type
the project prefix and then get a dynamic menu of all commands in the keymap.
Also, I'm free to re-define the project prefix key, e.g. I bound
`project-prefix-map` to `SPC p` in evil-mode, but the user could also bind it to
`C-c p` for all commands if he wanted to.

The downside is that by binding to `project-prefix-map` directly, we are
overriding the keybindings for some built-in commands like `project-find-file`.
Is this why you used `C-c p` as a prefix in the example? Anyway I understand
this is opinionated, so feel free to reject the PR if you want to use a
different prefix for consult-project-extra commands.